fix flakiness
