+++ zitadel/v2/org/remote_test.go
@@ -1,0 +1,29 @@
+package org_test
+
+import (
+	"fmt"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+	"github.com/zitadel/terraform-provider-zitadel/zitadel/v2/helper/test_utils"
+	"github.com/zitadel/zitadel-go/v2/pkg/client/zitadel/admin"
+	"github.com/zitadel/zitadel-go/v2/pkg/client/zitadel/org"
+)
+
+func checkRemoteProperty(frame *test_utils.OrgTestFrame) func(string) resource.TestCheckFunc {
+	return func(expect string) resource.TestCheckFunc {
+		return func(state *terraform.State) error {
+			remoteResource, err := frame.Admin.GetOrgByID(frame, &admin.GetOrgByIDRequest{Id: frame.State(state).ID})
+			if err != nil {
+				return err
+			}
+			actual := remoteResource.GetOrg().GetName()
+			if remoteResource.GetOrg().GetState() == org.OrgState_ORG_STATE_REMOVED {
+				return fmt.Errorf("org is removed: %w", test_utils.ErrNotFound)
+			}
+			if actual != expect {
+				return fmt.Errorf("expected %s, but got %s", expect, actual)
+			}
+			return nil
+		}
+	}
+}
